extract YOLO-based detection class
