+++ uavf_2024/imaging/detection/__init__.py
@@ -1,0 +1,2 @@
+from .shape_detector import ShapeDetector
+from .one_detector import OneDetector
