+++ uavf_2024/imaging/detection/one_detector.py
@@ -1,0 +1,5 @@
+from .yolo_detector import YOLODetector
+
+
+class OneDetector(YOLODetector):
+    ...
