--- uavf_2024/imaging/shape_detection/__init__.py
@@ -1,1 +1,0 @@
-from .shape_detector import ShapeDetector
